+++ infra/tf/better_uptime/outputs.tf
@@ -1,0 +1,5 @@
+output "status_page_domain" {
+  value = betteruptime_status_page.status_page.subdomain
+  description = "The domain of the Better Uptime status page"
+}
+