+++ infra/tf/better_uptime/providers.tf
@@ -1,0 +1,8 @@
+terraform {
+	required_providers {
+		betteruptime = {
+			source = "BetterStackHQ/better-uptime"
+			version = "0.5.1"
+		}
+	}
+}