+++ infra/tf/better_uptime/vars.tf
@@ -1,0 +1,20 @@
+# MARK: Better Uptime
+variable "better_uptime_groups" {
+	type = list(object({
+		id = string
+		name = string
+		monitors = list(object({
+			id = string
+			url = string
+			public_name = string
+		}))
+	}))
+}
+
+variable "better_uptime" {
+	type = object({
+		company_name = string
+		company_url = string
+		company_subdomain = string
+	})
+}