+++ infra/tf/k8s_cluster_k3d/main.tf
@@ -72,6 +72,14 @@ resource "k3d_cluster" "main" {
 
 	}
 
+	# kubectl
+	port {
+		host = "0.0.0.0"
+		host_port = 6443
+		container_port = 6443
+		node_filters = ["server:0"]
+	}
+
 	k3s {
 		extra_args {
 			arg = "--disable=traefik"
@@ -85,3 +93,15 @@ resource "k3d_cluster" "main" {
 	}
 }
 
+resource "null_resource" "post_cluster_creation" {
+	depends_on = [k3d_cluster.main]
+
+	provisioner "local-exec" {
+		command = <<EOF
+			until docker ps | grep -q k3d-${k3d_cluster.main.name}-server-0; do
+				sleep 1
+			done
+			docker exec k3d-${k3d_cluster.main.name}-server-0 mount --make-rshared /
+		EOF
+	}
+}