+++ lib/bolt/core/src/tasks/infra/mod.rs
@@ -275,6 +275,17 @@ pub fn build_plan(
 		}
 	}
 
+	// BetterUptime
+	if ctx.ns().better_uptime.is_some() {
+		plan.push(PlanStep {
+			name_id: "better_uptime",
+			kind: PlanStepKind::Terraform {
+				plan_id: "better_uptime".into(),
+				needs_destroy: true,
+			},
+		});
+	}
+
 	// S3
 	let s3_providers = &ctx.ns().s3.providers;
 	if s3_providers.minio.is_some() {