+++ svc/api/status/src/route/matchmaker.rs
@@ -1,7 +1,10 @@
 use api_helper::{anchor::WatchIndexQuery, ctx::Ctx};
 use proto::backend::pkg::*;
+use rivet_api::{
+	apis::{configuration::Configuration, *},
+	models,
+};
 use rivet_operation::prelude::*;
-use rivet_status_server::models;
 use serde::{Deserialize, Serialize};
 
 use crate::auth::Auth;
@@ -17,16 +20,62 @@ pub async fn status(
 	ctx: Ctx<Auth>,
 	_watch_index: WatchIndexQuery,
 	query: StatusQuery,
-) -> GlobalResult<models::MatchmakerResponse> {
+) -> GlobalResult<serde_json::Value> {
 	let domain_cdn = unwrap!(util::env::domain_cdn());
 
-	// Build client
-	let client = rivet_matchmaker::Config::builder()
-		.set_uri("http://traefik.traefik.svc.cluster.local:80/matchmaker")
-		.build_client();
+	// Find namespace ID
+	let game_res = op!([ctx] game_resolve_name_id {
+		name_ids: vec!["sandbox".into()],
+	})
+	.await?;
+	let game_id = unwrap_with!(
+		game_res.games.first().and_then(|x| x.game_id),
+		INTERNAL_STATUS_CHECK_FAILED,
+		error = "missing sandbox game"
+	);
+	let ns_resolve = op!([ctx] game_namespace_resolve_name_id {
+		game_id: Some(game_id),
+		name_ids: vec!["prod".into()],
+	})
+	.await?;
+	let ns_id = unwrap_with!(
+		ns_resolve.namespaces.first().and_then(|x| x.namespace_id),
+		INTERNAL_STATUS_CHECK_FAILED,
+		error = "missing prod namespace"
+	);
 
 	// Create bypass token
-	let token_res = op!([ctx] token_create {
+	let ns_token_res = op!([ctx] token_create {
+		token_config: Some(token::create::request::TokenConfig {
+			ttl: util::duration::minutes(15),
+		}),
+		refresh_token_config: None,
+		issuer: "api-status".to_owned(),
+		client: Some(ctx.client_info()),
+		kind: Some(token::create::request::Kind::New(token::create::request::KindNew {
+			entitlements: vec![
+				proto::claims::Entitlement {
+					kind: Some(proto::claims::entitlement::Kind::GameNamespacePublic(
+						proto::claims::entitlement::GameNamespacePublic {
+							namespace_id: Some(ns_id),
+						},
+					))
+				},
+				proto::claims::Entitlement {
+					kind: Some(
+						proto::claims::entitlement::Kind::Bypass(proto::claims::entitlement::Bypass {})
+					)
+				}
+			],
+		})),
+		label: Some("ns_pub".to_owned()),
+		..Default::default()
+	})
+	.await?;
+	let ns_token = unwrap_ref!(ns_token_res.token).token.clone();
+
+	// Bypass token
+	let bypass_token_res = op!([ctx] token_create {
 		token_config: Some(token::create::request::TokenConfig {
 			ttl: util::duration::minutes(15),
 		}),
@@ -37,7 +86,7 @@ pub async fn status(
 			entitlements: vec![
 				proto::claims::Entitlement {
 					kind: Some(
-						proto::claims::entitlement::Kind::Bypass(proto::claims::entitlement::Bypass { })
+						proto::claims::entitlement::Kind::Bypass(proto::claims::entitlement::Bypass {})
 					)
 				}
 			],
@@ -46,20 +95,76 @@ pub async fn status(
 		..Default::default()
 	})
 	.await?;
-	let token = unwrap_ref!(token_res.token).token.clone();
+	let bypass_token = unwrap_ref!(bypass_token_res.token).token.clone();
+
+	// Create client
+	let mut headers = reqwest::header::HeaderMap::new();
+	headers.insert("host", util::env::host_api().parse()?);
+	headers.insert(
+		"cf-connecting-ip",
+		reqwest::header::HeaderValue::from_str("127.0.0.1")?,
+	);
+	headers.insert(
+		"x-coords",
+		reqwest::header::HeaderValue::from_str("0.0,0.0")?,
+	);
+	headers.insert(
+		"x-bypass-token",
+		reqwest::header::HeaderValue::from_str(&bypass_token)?,
+	);
+
+	let client = reqwest::ClientBuilder::new()
+		.default_headers(headers)
+		.build()?;
+	let config = Configuration {
+		base_path: "http://traefik.traefik.svc.cluster.local:80".into(),
+		bearer_access_token: Some(ns_token),
+		client,
+		..Default::default()
+	};
 
 	tracing::info!("finding lobby");
-	let origin = format!("https://sandbox.{domain_cdn}/");
-	client
-		.find_lobby()
-		.origin(origin)
-		.bypass_token(token)
-		.game_modes("default")
-		.regions(&query.region)
-		.send()
-		.await?;
+	let res = matchmaker_lobbies_api::matchmaker_lobbies_create(
+		&config,
+		models::MatchmakerLobbiesCreateRequest {
+			game_mode: "custom".into(),
+			region: Some(query.region.clone()),
+			..Default::default()
+		},
+	)
+	.await;
+	let res = match res {
+		Ok(x) => x,
+		Err(err) => {
+			bail_with!(
+				INTERNAL_STATUS_CHECK_FAILED,
+				error = format!("find lobby: {:?}", err)
+			)
+		}
+	};
 
-	// TODO: Include connecting to socket, see stress test
+	// Make HTTP request through the socket
+	let port_default = unwrap!(res.lobby.ports.get("default"));
+	let port_host = unwrap_ref!(port_default.host);
+	let res = reqwest::get(format!("https://{port_host}/health")).await;
+	let res = match res {
+		Ok(x) => x,
+		Err(err) => {
+			bail_with!(
+				INTERNAL_STATUS_CHECK_FAILED,
+				error = format!("connect to lobby: {:?}", err)
+			)
+		}
+	};
+	let res = match res.error_for_status() {
+		Ok(x) => x,
+		Err(err) => {
+			bail_with!(
+				INTERNAL_STATUS_CHECK_FAILED,
+				error = format!("connect to lobby status: {:?}", err)
+			)
+		}
+	};
 
-	Ok(models::MatchmakerResponse {})
+	Ok(serde_json::json!({}))
 }